Optimize await, fix controller
- controller wasn't using DI instance across the board yet
- improve the await
