+++ csharp/src/Bargreen.Services/Bargreen.Services.csproj
@@ -4,8 +4,4 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="Implementations" />
-  </ItemGroup>
-
 </Project>
